Remove the logstash service reference from the docker-compose
The service definition here was trying to add volumes to logstash
without the service context.
This firstly isn't possible since no image was referenced - secondly it
was unnecessary as Logstash as part of instant openhie already has those
volumes attached. In this scenario all we need to do is add the files to
the config and restart logstash
The lostash helper might not even be needed...

DIO-106